+++ src/api/event/AdEvent.ts
@@ -111,11 +111,17 @@ export enum AdEventType {
 
   /**
    * Dispatched when an ad is clicked.
+   *
+   * @remarks
+   * <br/> - Available only on Android.
    */
   AD_CLICKED = 'adclicked',
 
   /**
    * Dispatched when a non-clickthrough portion of an ad is tapped.
+   *
+   * @remarks
+   * <br/> - Available only on Android.
    */
   AD_TAPPED = 'adtapped',
 
@@ -125,7 +131,9 @@ export enum AdEventType {
    * - On iOS and Android mobile apps, the SDK will navigate to the landing page.
    * - On tvOS and Android TV, the SDK will present a modal dialog containing the VAST icon fallback image.
    *
-   * @see <a href="https://developers.google.com/ad-manager/dynamic-ad-insertion/sdk/tvos/reference/Enums/IMAAdEventType#kimaadevent_icon_tapped">iOS IMA reference</a>
+   * @remarks
+   * <br/> - Available only on Android.
+   *
    * @see <a href="https://developers.google.com/interactive-media-ads/docs/sdks/android/client-side/api/reference/com/google/ads/interactivemedia/v3/api/AdEvent.AdEventType#public-static-final-adevent.adeventtype-icon_tapped">Android IMA reference</a>
    */
   AD_ICON_TAPPED = 'adicontapped',
@@ -133,7 +141,9 @@ export enum AdEventType {
   /**
    * Dispatched when the user has closed the icon fallback image dialog.
    *
-   * @see <a href="https://developers.google.com/ad-manager/dynamic-ad-insertion/sdk/tvos/reference/Enums/IMAAdEventType#kimaadevent_icon_fallback_image_closed">iOS IMA reference</a>
+   * @remarks
+   * <br/> - Available only on Android.
+   *
    * @see <a href="https://developers.google.com/interactive-media-ads/docs/sdks/android/client-side/api/reference/com/google/ads/interactivemedia/v3/api/AdEvent.AdEventType#public-static-final-adevent.adeventtype-icon_fallback_image_closed">Android IMA reference</a>
    */
   AD_ICON_FALLBACK_IMAGE_CLOSED = 'adiconfallbackimageclosed'