+++ src/api/event/AdEvent.ts
@@ -108,4 +108,43 @@ export enum AdEventType {
    * Dispatched when the ad has stalled playback to buffer.
    */
   AD_BUFFERING = 'adbuffering',
+
+  /**
+   * Dispatched when an ad is clicked.
+   *
+   * @remarks
+   * <br/> - Available only on Android.
+   */
+  AD_CLICKED = 'adclicked',
+
+  /**
+   * Dispatched when a non-clickthrough portion of an ad is tapped.
+   *
+   * @remarks
+   * <br/> - Available only on Android.
+   */
+  AD_TAPPED = 'adtapped',
+
+  /**
+   * Dispatched when the user has tapped an ad icon, for example the 'Why this ad' (WTA) icon.
+   *
+   * - On iOS and Android mobile apps, the SDK will navigate to the landing page.
+   * - On tvOS and Android TV, the SDK will present a modal dialog containing the VAST icon fallback image.
+   *
+   * @remarks
+   * <br/> - Available only on Android.
+   *
+   * @see <a href="https://developers.google.com/interactive-media-ads/docs/sdks/android/client-side/api/reference/com/google/ads/interactivemedia/v3/api/AdEvent.AdEventType#public-static-final-adevent.adeventtype-icon_tapped">Android IMA reference</a>
+   */
+  AD_ICON_TAPPED = 'adicontapped',
+
+  /**
+   * Dispatched when the user has closed the icon fallback image dialog.
+   *
+   * @remarks
+   * <br/> - Available only on Android.
+   *
+   * @see <a href="https://developers.google.com/interactive-media-ads/docs/sdks/android/client-side/api/reference/com/google/ads/interactivemedia/v3/api/AdEvent.AdEventType#public-static-final-adevent.adeventtype-icon_fallback_image_closed">Android IMA reference</a>
+   */
+  AD_ICON_FALLBACK_IMAGE_CLOSED = 'adiconfallbackimageclosed'
 }