+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `resize` player event enabling monitoring of player resizes.
 - Added `width` and `height` properties to `THEOplayer` containing the player's current dimensions in pixels.
 - Added the possibility to set the player volume on iOS programmatically via the player API.
-- Added `AD_CLICKED`, `AD_TAPPED`, `AD_ICON_TAPPED` and `AD_ICON_FALLBACK_IMAGE_CLOSED` ad events for mobile platforms.
+- Added `AD_CLICKED`, `AD_TAPPED`, `AD_ICON_TAPPED` and `AD_ICON_FALLBACK_IMAGE_CLOSED` ad events for Android platforms.
 
 ### Fixed